schemeshard: align DbRefCount layer filenames with renamed types
Follow-up to the type rename:
- schemeshard_self_ref_map.h -> schemeshard_db_ref_map.h   (TDbRefMap)
- schemeshard_path_ref.{h,cpp} -> schemeshard_path_db_ref.{h,cpp}  (TPathDbRef)
- ut_base/ut_self_ref_map.cpp -> ut_base/ut_db_ref_map.cpp

Include sites and ya.make updated. No behavior change; ut_base + ut_ttl 298/298 GOOD.
